bleframe moved to views folder, shock widget to widgets
